PythonRpcClient/requirements.txt
grpcio==1.41.1
grpcio-tools==1.41.1
pip==21.1.1
protobuf==3.19.1
setuptools==56.0.0
six==1.16.0
